Кошик 4

+++ apps/cart/views.py
@@ -27,7 +27,7 @@ def cart_add(request, product_id):
         product_id_str = str(product_id)
         cart_item = cart.cart.get(product_id_str, {})
         item_quantity = cart_item.get('quantity', 0)
-        item_price = Decimal(cart_item.get('price', product.get_price_for_user(cart.user, item_quantity)))
+        item_price = float(cart_item.get('price', product.get_price_for_user(cart.user, item_quantity)))
         
         return JsonResponse({
             'success': True,
@@ -40,8 +40,8 @@ def cart_add(request, product_id):
             'item': {
                 'product_id': product_id,
                 'quantity': item_quantity,
-                'price': float(item_price),
-                'total': float(item_price * item_quantity),
+                'price': item_price,
+                'total': item_price * item_quantity,
             }
         })
